pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["prometheus*"]

[project]
name = "prometheus"
version = "0.1.0"
description = "Prometheus – Quantitative Trading System (thin layer, imports apathis for info-layer)"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "apathis @ file:///home/feanor/coding/apathis",
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "psycopg2-binary>=2.9.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.12.0",
  "python-dateutil>=2.8.0",
  "pyyaml>=6.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "ib_insync>=0.9.86",
  "scipy>=1.10.0",
  "scikit-learn>=1.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
